+++ hugr/src/std_extensions/arithmetic/int_ops/const_fold/test.rs
